rewrite Action/Fluent/Function/Reference init
Actions, Fluents and Functions are not Statements or Expressions any
more. They just expose their Expression type class as a typedef, and
References to them inherit from that typedef.

Declaring and defining Actions, Fluents and Functions now has a more
clearly defined semantics. This also fixes the use after free bugs that
were present in the previous implementation.

The way References are parsed and initialized is changed to allow for
nested References of arbitrary depth.