redesign model initialization for up traversal
Every Expression now has a pointer to its parent. This allows walking
the expression tree up from any element, not just down. It is required
for translating functions to readylog since the function/3 predicate
needs to declare a return variable that is then set anywhere in the
body. To translate golog++ return statements, we need to walk up the
expression tree to the function semantics implementation that declares
the return variable.