modularize parser compilation units
Having the entire grammar in one huge header file is a Bad Idea. This
commit splits it up into small units that can be compiled in parallel,
speeding up the process. Individual units also use much less RAM than
the monolithic header to compile, so machines with 4 GB or less
shouldn't crash any more ;-)

At the same time, the pick() statement is changed to allow specification
of domains, which is required for ReadyLog.