model, semantics: implement ExogActions
This is a major redesign of how parameterized (i.e. executed and
exogenous) actions are modeled. This commit introduces a new class
template Grounding<T>, which represents a reference to a parameterizable
element (i.e. a Global) with all arguments bound to constant values.
Grounding<Action> resp. Grounding<ExogAction> is then the base class for
every Activity, Transition and ExogEvent.