model: remove compile-time type system
Representing expression types as template arguments was an attempt to
enforce strict expression type consistency in the API. However it turns
out that it's impossible to enforce everywhere and it adds a lot of
complexity. So since we need runtime type checking anyways, this commit
is the first move towards doing *all* type checking at runtime. This
allows us to turn many previously templated classes into regular
classes, which should eventually simplify the parser as well.