parser: Adapt to new runtime type system
This amounts to a rewrite of large portions of the parser. Basically
everything that differentiates types needed to be touched and adapted to
not handle types at compile time. It's a huge simplification that
eliminates a bunch of 5-way branches in the very middle of the syntax
graph, which should hopefully reduce complexity a lot. The code also
looks much less verbose without all the templates.