readylog: implement list semantics
including all operations. This also changes the CompoundType semantics
so that both lists and compound values store their type name in
readylog, as well. This is needed when values are translated back from
prolog to golog++ (e.g. on a transition), because we need to know the
type to construct a golog++ Value.