allow general initially: statements
Before, an initially: statement could only define initial values for
concrete args, and the value also needed to be an actual value of the
appropriate type.

Now the arguments of initial fluent values can be unbound, and the value
can be "null", allowing the user to define "catch-all" initial values.
It should also be possible to mix bound and unbound arguments.