split Semantics class model
We now have two templated class hierarchies: AbstractSemantics<T> and
Semantics<T>. The AbstractSemantics represent everything that is
independent of a particular implementation, and can thus be used in the
model. The Semantics hierarchy is the actual implementation, and it is
undefined within the model. The reason for introducing this is to allow
a generic ExecutionContext to evaluate complex value expressions
individually, e.g. as part of an action mapping. While the whole mapping
logic could also be implemented within the semantics, that would make it
specific to a certain semantics, and that is not what we're trying to
achieve here.