fix unsigned int Value ReadyLog semantics
Wrapping the boost::variant get into cast operators on the Value class
was a borked design decision, since it doesn't trigger intermediate
conversion if for upcasts to types that could also be contained in the
variant. For details see the TODO. The gist of it is that the cast-based
Value conversion API was a stupid design decision.