parser: fix member access
Get rid of the stupid differentiation between mixed list access and
mixed field access. Most of the time we don't care anyways (we only care
about the final type of the expression, not whether it comes out of a
list or a compound). And the few times we do care (EffectAxiom and
Assignment), can be solved by a dynamic_cast. On top of that, we might
even ditch the differentiation there, as well.