model: introduce Instruction class
This splits non-evaluatable language elements off into a class hierarchy
separate from Expressions. Evaluatable Expressions and Transitionable
Instructions are completely disjunct in their functionality, use and
syntax.

This is a significant change deep in the golog++ class model, and as
such it entails major redesign of some aspects. Specifically, the syntax
of functions now reflects their truly functional nature in that they
don't use the block notation with curly braces any more.

This of course also leads to major changes to the parser and the
semantics, which follow in the next commits.