semantics: make progression safer
There's a potential BAD_TERM in the progression function where we pull
the new history out of it. So far nothing has happened, but that could
also mean it's just very unlikely to fail because it's immediately
converted to a ManagedWord. Still, it's probably better to use an EC_ref
instead of an EC_word just to make sure.