make reference argument parsing more deterministic
We were basically parsing each reference argument blindly, i.e. try
parsing each argument as each of the predefined types, and then see
whether the results match the signature of the referee.

Turns out that this is one of the most expesive aspects of the whole
grammar since references can come up anywhere, in particular as
arguments to references (leading to an exponential blowup).

This commit now resolves the reference immediately (i.e. before any of
the arguments are parsed) and then dynamically picks the right parser
for each argument type, which should significantly reduce complexity,
in particular when we have nested references.

However, with this we have to reduce the identity of a referee to just
its name, instead of name/arity as before. So we can't have e.g. a
function called bla(number x) AND a fluent called bla(number x, string
y) in the same program.