semantics/readylog: move history helpers to execution
Helper functions to retrieve gologpp data from prolog terms may be
useful when handling the execution in order to synthesize plans from
terms. Therefore, move them to become static members of the
ReadyLogContext class.

Also, refactor the functionality to get a transition from an action
term to be a separate function.