fix design error in semantics interface
The readylog semantics cannot define a general Semantics<ModelElement>
because platform model elements are ModelElements, too, and they have to
be able to define their own semantics. This entails changes to the code
model's semantics interface as well, because the assumption that there
can be a general Semantics<ModelElement> was reflected there, too.