readylog: remodel variable initialization
This is now the sledgehammer approach I've been trying to avoid, but
turns out it matches eclipse-clp's behavior most closely: Every single
variable is marked "dead" after an EC_resume (which it actually is by
then), and dead variables get a fresh term assigned which makes them not
dead. This should be robust enough, doing away with the error-prone
approach of initializing variables individually while traversing the
code model.