don't mess with time points out of transformation
When appending a timed action to the plan and it's an end(...) action,
its time window gets adjusted to account for the action's duration IF a
matching start(...) action is found in the plan. However we don't want
to do this on the timed actions we get out of the plan.