execution: fix initial plan time windows
This removes some unnecessary waits, but the taptenc transformation
still adds some because it needs to guarantee model consistency just by
setting time windows. This means that constrained actions are always
delayed until the moment the constraints are guaranteed to be satisfied
or the platform model is violated and a re-transformation happens.