implement exog_function evaluation from readylog
This now allows us to call exogenous functions from anywhere in the
code, not just in sensing assignments.

For this we had to sacrifice the mapping functionality, so
exog_functions calls are always passed to the backend with their name
and argument names as they are. However that shouldn't be a problem
since we can now simply wrap them in regular functions.