platform: always transform plan at t=0
The transformation interprets non-zero plan start times as waiting
periods because it doesn't (and isn't supposed to) have a concept of the
current time. So we need to shift all action time windows back so the
plan always starts at t=0. Additionally, make sure the first action's
time window is always larger than 0.