model: redesign event architecture
The Reference< > template gained a new parameter for the argument type.
It can be an arbitrary Expression or a Value. There is also a new,
specific event type for use with the exog queue, which makes use of
this new Reference<T, Value> (i.e. Grounding<T>).

This change may cause some minor breakage in downstream code, but it
should be straightforward to fix because the change is making compiler
errors more specific and adds some additional type safety around the
handling of events.