Fix on_cancel construct
The `CANCELLED` state of an activity was initially named `PREEMPTED`. In the `to_string(Activity::State)` function the rename was missed, but it is actually used in the ReadyLog semantics to generate the code for the `on_cancel ...` branch of the `during` construct. So we ended up with ReadyLog code like `if (state(A) = preempted, BLABLOCK)`, causing `BLABLOCK` to never be executed because `state(A) = cancelled`.